Fixed One to Many relationship and tidy up ORM

+++ src/main/java/com/net/SprintOne/repositories/RoleRepository.java
@@ -2,9 +2,10 @@ package com.net.SprintOne.repositories;
 
 import com.net.SprintOne.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
     List<Role> findByName(String name);
 }
